Edit and Delete Stream Protection Commit:
Add basic logic to prevent unauthorized deletion or edits of streams.

+++ src/components/streams/StreamDelete.js
@@ -4,20 +4,24 @@ import Button from "../ui-elements/Button";
 import history from "../../history";
 import {connect} from "react-redux";
 import {deleteStream, fetchStream} from "../../redux/actions";
-import {Link} from "react-router-dom";
 
 class StreamDelete extends React.Component {
+
+  componentDidMount() {
+    this.props.fetchStream(this.props.match.params.id);
+  }
 
   onDismissHandler = () => {
     history.push("/");
   };
 
   onDeleteHandler = () => {
-    this.props.deleteStream(this.props.match.params.id);
-    history.push("/");
+    if (this.props.auth.userId === this.props.stream.userId) {
+      this.props.deleteStream(this.props.match.params.id);
+    }
   };
 
-  actions = () => (
+  renderActions = () => (
     <React.Fragment>
       <Button
         buttonType="button-animated"
@@ -26,29 +30,41 @@ class StreamDelete extends React.Component {
         buttonIcon="trash"
         onClick={this.onDeleteHandler}
       />
-      <Link to={"/"}>
-        <Button
-          buttonType="button-animated"
-          buttonClass="ui button basic"
-          buttonText="Cancel"
-          buttonIcon="right angle"
-        />
-      </Link>
+      <Button
+        buttonType="button-animated"
+        buttonClass="ui button basic"
+        buttonText="Cancel"
+        buttonIcon="right angle"
+        onClick={this.onDismissHandler}
+      />
     </React.Fragment>
   );
 
+  renderContent = () => {
+    if (!this.props.stream) {
+      return "Are you sure you want to delete this stream?"
+    } else {
+      return `Are you sure you want to delete the stream with title: ${this.props.stream.title}`
+    }
+  };
+
   render() {
     return (
-      <div>
         <Modal
           title={"Delete Stream"}
-          content={"Are you sure you want to delete this stream?"}
-          actions={this.actions()}
+          content={this.renderContent()}
+          actions={this.renderActions()}
           onDismiss={this.onDismissHandler}
         />
-      </div>
     );
   };
 }
 
-export default connect(null, {deleteStream, fetchStream})(StreamDelete);
+const mapStateToProps = (state, ownProps) => {
+  return {
+    stream: state.streams[ownProps.match.params.id],
+    auth: state.auth
+  }
+};
+
+export default connect(mapStateToProps, {deleteStream, fetchStream})(StreamDelete);
